Provide TEdgeConfig as wrapper around TEdgeConfigDto for now
