+++ tedge_config/src/lib.rs
@@ -7,8 +7,8 @@ mod types;
 
 pub use self::config_setting::*;
 pub use self::error::*;
+pub use self::providers::tedge_config::*;
 pub use self::providers::tedge_config_dto::*;
 pub use self::settings::*;
 pub use self::traits::*;
 pub use self::types::*;
-pub use TEdgeConfigDto as TEdgeConfig;
